+++ kwcoco/__init__.py
@@ -269,7 +269,7 @@ Testing:
 
 """
 
-__version__ = '0.4.0'
+__version__ = '0.4.1'
 
 
 __submodules__ = {